test(fleet): split the sightings tests out from under the 500-line cap
My two new tests took `fleet/tests.rs` to 526 lines and CI caught it. My own
pre-commit check reported zero over-cap files because I ran it on a different
branch — the check was right, the branch it looked at was not.

The seam is real rather than a line count: `fleet/tests.rs` is about what the
fleet view REPORTS, and these are about what an unauthenticated network can make
it STORE.
